moved action points categories into public schema
