+++ src/etools/applications/action_points/categories/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = __name__ + '.apps.AppConfig'
