+++ src/etools/applications/action_points/categories/apps.py
@@ -1,0 +1,7 @@
+
+from django.apps import AppConfig as BaseAppConfig
+
+
+class AppConfig(BaseAppConfig):
+    name = __name__.rpartition('.')[0]
+    verbose_name = 'Action Points Categories'
