+++ src/etools/applications/action_points/models.py
@@ -11,6 +11,7 @@ from model_utils.models import TimeStampedModel
 from ordered_model.models import OrderedModel
 from unicef_snapshot.models import Activity
 
+from etools.applications.action_points.categories.models import Category
 from etools.applications.action_points.transitions.conditions import ActionPointCompleteActionsTakenCheck
 from etools.applications.EquiTrack.utils import get_environment
 from etools.applications.notification.models import Notification
@@ -19,32 +20,8 @@ from etools.applications.utils.common.urlresolvers import build_frontend_url
 from etools.applications.utils.groups.wrappers import GroupWrapper
 
 
-class Category(OrderedModel, models.Model):
-    MODULE_CHOICES = Choices(
-        ('apd', _('Action Points')),
-        ('t2f', _('Trip Management')),
-        ('tpm', 'Third Party Monitoring'),
-        ('audit', _('Financial Assurance')),
-    )
-
-    module = models.CharField(max_length=10, choices=MODULE_CHOICES, verbose_name=_('Module'))
-    description = models.TextField(verbose_name=_('Description'))
-
-    class Meta:
-        unique_together = ("description", "module", )
-        ordering = ('module', 'order')
-
-    def __str__(self):
-        return '{}: {}'.format(self.module, self.description)
-
-
 class ActionPoint(TimeStampedModel):
-    MODULE_CHOICES = Choices(
-        ('apd', _('Action Points')),
-        ('t2f', _('Trip Management')),
-        ('tpm', _('Third Party Monitoring')),
-        ('audit', _('Financial Assurance')),
-    )
+    MODULE_CHOICES = Category.MODULE_CHOICES
 
     STATUSES = Choices(
         ('open', _('Open')),
