+++ src/etools/applications/action_points/serializers.py
@@ -7,10 +7,11 @@ from unicef_snapshot.models import Activity
 from unicef_snapshot.serializers import SnapshotModelSerializer
 
 from etools.applications.EquiTrack.utils import get_current_site
-from etools.applications.action_points.models import ActionPoint, Category
+from etools.applications.action_points.categories.serializers import CategorySerializer
+from etools.applications.action_points.models import ActionPoint
+from etools.applications.action_points.categories.models import Category
 from unicef_locations.serializers import LocationLightSerializer
 
-from etools.applications.attachments.serializers_fields import ModelChoiceField
 from etools.applications.partners.serializers.interventions_v2 import BaseInterventionListSerializer
 from etools.applications.partners.serializers.partner_organization_v2 import MinimalPartnerOrganizationListSerializer
 from etools.applications.permissions2.serializers import PermissionsBasedSerializerMixin
@@ -20,17 +21,6 @@ from etools.applications.users.serializers_v3 import MinimalUserSerializer
 from etools.applications.utils.common.serializers.fields import SeparatedReadWriteField
 from etools.applications.utils.common.serializers.mixins import UserContextSerializerMixin
 from etools.applications.utils.writable_serializers.serializers import WritableNestedSerializerMixin
-
-
-class CategoryModelChoiceField(ModelChoiceField):
-    def get_choice(self, obj):
-        return obj.id, obj.description
-
-
-class CategorySerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Category
-        fields = ('id', 'module', 'description')
 
 
 class ActionPointBaseSerializer(UserContextSerializerMixin, SnapshotModelSerializer, serializers.ModelSerializer):
