+++ src/etools/applications/action_points/tests/factories.py
@@ -7,7 +7,8 @@ import factory.fuzzy
 from django_comments.models import Comment
 
 from etools.applications.EquiTrack.utils import get_current_site
-from etools.applications.action_points.models import ActionPoint, Category
+from etools.applications.action_points.models import ActionPoint
+from etools.applications.action_points.categories.models import Category
 from etools.applications.firms.tests.factories import BaseUserFactory
 from unicef_locations.tests.factories import LocationFactory
 from etools.applications.partners.tests.factories import InterventionFactory, ResultFactory
