+++ src/etools/applications/action_points/views.py
@@ -13,9 +13,11 @@ from etools.applications.action_points.export.renderers import ActionPointCSVRen
 from etools.applications.action_points.export.serializers import ActionPointExportSerializer
 from etools.applications.action_points.filters import ReferenceNumberOrderingFilter, RelatedModuleFilter
 from etools.applications.action_points.metadata import ActionPointMetadata
-from etools.applications.action_points.models import ActionPoint, Category
+from etools.applications.action_points.models import ActionPoint
+from etools.applications.action_points.categories.models import Category
 from etools.applications.action_points.serializers import ActionPointListSerializer, ActionPointSerializer, \
-    ActionPointCreateSerializer, CategorySerializer
+    ActionPointCreateSerializer
+from etools.applications.action_points.categories.serializers import CategorySerializer
 from etools.applications.permissions2.conditions import ObjectStatusCondition
 from etools.applications.permissions2.views import PermittedFSMActionMixin, PermittedSerializerMixin
 from etools.applications.utils.common.pagination import DynamicPageNumberPagination
