+++ src/etools/applications/audit/serializers/engagement.py
@@ -3,9 +3,9 @@ from copy import copy
 from django.utils.translation import ugettext as _
 from rest_framework import serializers
 
-from etools.applications.action_points.models import Category
-from etools.applications.action_points.serializers import ActionPointBaseSerializer, HistorySerializer, \
-    CategoryModelChoiceField
+from etools.applications.action_points.categories.models import Category
+from etools.applications.action_points.serializers import ActionPointBaseSerializer, HistorySerializer
+from etools.applications.action_points.categories.serializers import CategoryModelChoiceField
 from etools.applications.attachments.models import FileType
 from etools.applications.attachments.serializers import BaseAttachmentSerializer
 from etools.applications.attachments.serializers_fields import FileTypeModelChoiceField
